Initial commit. First working version of Gesture Module as an Fennec Extension.
The Gesture Module is at the same iteration than the first patch posted on bugzilla, plus simple event raising for detected gestures.
The user-experience prototyping has some simple demonstrations of listening to gesture events to perform actions such as opening or closing a tab, adding a page to the bookmarks or opening the bookmarks editor.