Added API for acessing the Gesture module and the latest movement data from global context.
Added the Gesture module as window.Gestures and a getter latestMovement for retrieving the latest movement (matched as a known gesture or not). This will be used later for the learning mode which should be able to get unrecognized gesture data.

This API is not final, and is a simplification to the alternative of creating a custom event type to carry the data since the default event cannot carry extra data. The alternative would be better for not having to add the Gestures module in global scope, but this is not bad too and is much simpler. And we might keep the Gestures module in global scope for other reasons (registering gestures, getting precise tracking info, etc.)