Algorithm improvement: min match value now changes depending on the movement length
This makes it support long movements much better, because otherwise long movements had to be much more precise. This change makes sense in an algorithm point of view: considering that we're calculating the edit distance, it is certain that longer gestures will be much harder to get exactly as their definition, and this change allows for more forgiveness on their detection.